weather_project/requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
joblib==1.4.2
nltk==3.8.1
regex==2024.5.15
sqlparse==0.5.0
tqdm==4.66.4
python-decouple==3.8
